[DEBUG] Handling of jsonc longer than ARDISCOVERY_CONNECTION_RX_BUFFER_SIZE
In previous version, a reading a json string made ox exaclty
ARDISCOVERY_CONNECTION_RX_BUFFER_SIZE characters (or any multiple of
this) would have resulted in an error:
For every 'complete' read, we reallocate the buffer, and do another
read. If this read fails, we discarded the entire data with
'ARDISCOVERY_ERROR_READ' error.

This commit add a check for a EAGAIN/EWOULDBLOCK read, which means
that we have no more data in the socket, and that the json string is
complete.

Change-Id: I95214d113853b84ff041f96dbdc6063f1f3ff803
Reviewed-on: https://gerrit.parrot.biz/20274
Reviewed-by: Marjorie Coulin <marjorie.coulin@parrot.com>
Reviewed-by: Nicolas Christe <nicolas.christe@parrot.com>