[BUGFIX 176707](Android) Proper flush of ARDiscovery device list
When the wifi quickly switch between two networks, we sometime receive
the disconnect event after the actual wifi reconnection. The previous
implementation relied on the wifiManager state to determine if the
event is a disconnection or a connection.

In such fast cases, the service was only seeing "connection" events.

This patch changes the way the service detects disconnections events,
by reading the actual intent extras.

Change-Id: I70c921a1d2e9764b61e20bb0e6946a2f11262977
Reviewed-on: https://gerrit.parrot.biz/21325
Reviewed-by: Nicolas Christe <nicolas.christe@parrot.com>