[DEV](DCIM) Add ARDISCOVERY_getProductFromPathName function
This function is needed by the DCIM-medias, as the product name is
only coded in the media thumbnail path.

This function will only compare the first N characters (where N is the
length of the product path) for each product, until a match is found.
This allow the caller to use a filename as an argument (e.g. passing
"Jumping_Night_something_someotherthing.jpg" will return the
JUMPING_NIGHT product).

Change-Id: Ieabc43056f9fe45308a32fa24dff34abd806ea4d
Reviewed-on: https://gerrit.parrot.biz/23164
Reviewed-by: Nicolas Christe <nicolas.christe@parrot.com>