[DEV] Declared a 2nd lib without EAccessory need in the atom
In order to provide a library that does not need to embed the
framework ExternalAccessory on iOS, a new libARDiscovery is
generated with all the code related to EAcc stubbed.

This library is name libardiscoverywithouteacc.

If no need to support Usb accessory (i.e SkyController2), this
new lib should be used.

Change-Id: I1695a43dcc11d12fca153b86d7441476541de881
Reviewed-on: https://gerrit.parrot.biz/48851
Reviewed-by: Jean-Baptiste Dubois <jean-baptiste.dubois@parrot.com>