[DEBUG] don't restart mux discovery sequence
Avoid restarting mux discovery: mux discovery is linked to the
mux lifecycle. Discovered device info are stored in USBAccessoryManager
and are notified when the USBAccessoryManager delegate is set

This allows bonjour discovery to simply unregister itself as USBAccessoryManager
delegate when stopped, and re-register itself when started.
If USBAccessoryManager did previously discover a device, this device
is immediately notified to the new registered delegate.

Change-Id: If8a3a8b62f19f281963fa8818427d3b552ecfb93
Reviewed-on: https://gerrit.parrot.biz/49649
Reviewed-by: Cyril HERVOUIN <cyril.hervouin@parrot.com>
Reviewed-by: Jean-Baptiste Dubois <jean-baptiste.dubois@parrot.com>
Reviewed-by: Djavan Bertrand <djavan.bertrand@parrot.com>