[DEV] (iOS) add USBAccessory proxy support
Add support to connect to a SkyController2 through "USB Debug Bridge" application (for
debugging). This configuration uses an application running on a Android device connected
to the sky controller 2 by USB that forward usb data on tcp/ip.

This allows to free iDevice usb connection for debugging.

USBAccessoryManager class is now a facade in front of a concrete implementation. By default
it uses USBAccessoryManagerEA class, that connects to the local iAP external accessory.

Implementation that connected to the USB accessory through "USB Debug Bridge" application
by wifi can be selected by specifying  an entry in application plist or a command line option.

Change-Id: Ia841161c333c9893679dce5bf68823dbb1fc1abc
Reviewed-on: https://gerrit.parrot.biz/50169
Reviewed-by: Djavan Bertrand <djavan.bertrand@parrot.com>