[DEV] Split eARDISCOVERY_PRODUCT enum
Add a new eARDISCOVERY_NETWORK_TYPE enum, filled by libARDiscovery, to
select the proper network type for a product.

This change allow a product to use multiple network types, and the
applciation to select the proper network type based on the discovery.

This will be required for SkyController_NG (SkyController with
SkyController 2 software) on Android, because it retains the legacy
WiFi connection, but adds the possibility to use an USB/libmux
connection.

Change-Id: I01e498666c1fa79cf7e970703b8514ff7130d809
Reviewed-on: https://gerrit.parrot.biz/56818
Reviewed-by: Djavan Bertrand <djavan.bertrand@parrot.com>