[DEV] : Add a way to tell discoveryService it should consider wifi available
On newer versions of android (>7), wifi may not be seen through ConnectivityManager
for exemple, if a mobile connection is already active. In which case, if the application
has established a wifi connection by another method (for exemple via bindProcessToNetwork
api), it can notify the discoveryService that wifi is present and accessible.

Change-Id: I4894da8172e27b3b597e68a87325efa6608110c2
Reviewed-on: https://gerrit.parrot.biz/61475
Reviewed-by: Djavan Bertrand <djavan.bertrand@parrot.com>